+++ Backend/Culinary-Assistant-Main/Culinary-Assistant.Core/ValueTypes/Nutrients.cs
@@ -9,10 +9,10 @@ namespace Culinary_Assistant.Core.ValueTypes
 {
 	public class Nutrients : ValueObject
 	{
-		public int Calories { get; private set; }
-		public int Proteins { get; private set; }
-		public int Fats { get; private set; }
-		public int Carbohydrates { get; private set; }
+		public double Calories { get; private set; }
+		public double Proteins { get; private set; }
+		public double Fats { get; private set; }
+		public double Carbohydrates { get; private set; }
 
 		protected override IEnumerable<object> GetEqualityComponents()
 		{
@@ -22,9 +22,9 @@ namespace Culinary_Assistant.Core.ValueTypes
 			yield return Carbohydrates;
 		}
 
-		public static Result<Nutrients> Create(int calories, int proteins, int fats, int carbohydrates)
+		public static Result<Nutrients> Create(double calories, double proteins, double fats, double carbohydrates)
 		{
-			List<int> nutrients = [calories, proteins, fats, carbohydrates];
+			List<double> nutrients = [calories, proteins, fats, carbohydrates];
 			if (nutrients.Any(n => n < 0))
 			{
 				return Result.Failure<Nutrients>("Ни один нутриент не может быть меньше 0");
